+++ app/layout.js
@@ -1,6 +1,6 @@
 import { GA_TRACKING_ID } from '../helpers/constants';
 import Script from 'next/script';
-import Navbar from './Navbar';
+import Navbar from '../components/Navbar';
 import DataContextProvider from '../providers/DataProvider';
 import { getData } from '../helpers/dataLoader';
